+++ CHANGELOG.rst
@@ -18,7 +18,8 @@ v1.0.0
 - Save WebElement in PageElement to avoid searching the same element multiple times
 - Refactor to rename get_element to get_web_element in Utils class and element to web_element in PageElement class
 - Allow to run API tests with behave: browser property must be empty
-- Baseline name property can contain *{platformVersion}* to add actual platform version to the baseline name
+- Baseline name property can contain *{PlatformVersion}* or *{RemoteNode}* to add actual platform version or remote
+  node name to the baseline name
 - Add force parameter to *assert_screenshot* methods to compare the screenshot even if visual testing is disabled by
   configuration